backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
python-multipart==0.0.6
pydantic==2.5.3
nltk==3.8.1
textblob==0.17.1
openai>=1.0.0
python-dotenv>=1.0.0
